requirements.txt
astroid==2.1.0
attrs==18.2.0
Automat==0.7.0
certifi==2018.11.29
chardet==3.0.4
constantly==15.1.0
HeapDict==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.0
PyHamcrest==1.9.0
pylint==2.2.2
pymongo==3.7.2
python-bittrex==0.3.0
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
six==1.12.0
toolz==0.9.0
tornado==5.1.1
urllib3==1.24.1
wrapt==1.10.11
zict==0.1.3
zope.interface==4.6.0
